requirements.txt
boto3==1.26.137
botocore==1.29.137
python-dateutil==2.8.2
email-validator==2.0.0
Pillow==10.4.0
python-multipart==0.0.6
PyJWT==2.8.0 